Tests/LibWeb/Text/expected/css/dynamic-viewport-overflow-hidden.txt

Content overflows viewport: true
Scroll width includes overflow: true
Horizontal wheel target: none
Scroll offset after wheel: 0
Scroll offset after scrollTo: 1
